+++ toolchain/check/BUILD
@@ -117,7 +117,7 @@ cc_library(
         "//common:vlog",
         "//toolchain/base:kind_switch",
         "//toolchain/base:pretty_stack_trace_function",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:format_providers",
         "//toolchain/lex:token_index",
@@ -241,6 +241,8 @@ cc_library(
         "//common:set",
         "//common:vlog",
         "//toolchain/base:index_base",
+        "//toolchain/base:value_ids",
+        "//toolchain/base:value_store",
         "//toolchain/sem_ir:file",
         "//toolchain/sem_ir:ids",
         "@llvm-project//llvm:Support",
