+++ toolchain/lex/BUILD
@@ -191,7 +191,7 @@ cc_library(
         ":tokenized_buffer",
         "//common:check",
         "//common:variant_helpers",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/source:source_buffer",
         "@llvm-project//llvm:Support",
@@ -223,7 +223,7 @@ cc_library(
         "//common:string_helpers",
         "//toolchain/base:index_base",
         "//toolchain/base:mem_usage",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/source:source_buffer",
         "@llvm-project//llvm:Support",
@@ -238,7 +238,7 @@ cc_library(
         ":lex",
         ":tokenized_buffer",
         "//common:check",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "@googletest//:gtest",
         "@llvm-project//llvm:Support",
     ],
@@ -255,7 +255,7 @@ cc_test(
         ":tokenized_buffer_test_helpers",
         "//testing/base:gtest_main",
         "//testing/base:test_raw_ostream",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:mocks",
         "//toolchain/testing:compile_helper",
@@ -274,7 +274,7 @@ cc_fuzz_test(
         ":lex",
         "//common:check",
         "//testing/fuzzing:libfuzzer_header",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:null_diagnostics",
         "@llvm-project//llvm:Support",
@@ -292,7 +292,7 @@ cc_binary(
         "//common:check",
         "//testing/base:benchmark_main",
         "//testing/base:source_gen_lib",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:null_diagnostics",
         "@abseil-cpp//absl/random",
