+++ toolchain/lex/lex.h
@@ -5,7 +5,7 @@
 #ifndef CARBON_TOOLCHAIN_LEX_LEX_H_
 #define CARBON_TOOLCHAIN_LEX_LEX_H_
 
-#include "toolchain/base/value_store.h"
+#include "toolchain/base/shared_value_stores.h"
 #include "toolchain/diagnostics/diagnostic_emitter.h"
 #include "toolchain/lex/tokenized_buffer.h"
 #include "toolchain/source/source_buffer.h"
