+++ toolchain/parse/BUILD
@@ -94,7 +94,7 @@ cc_library(
         "//common:check",
         "//common:ostream",
         "//toolchain/base:pretty_stack_trace_function",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:format_providers",
         "//toolchain/lex:token_kind",
@@ -127,6 +127,7 @@ cc_library(
         "//common:error",
         "//common:ostream",
         "//common:struct_reflection",
+        "//toolchain/base:value_store",
         "//toolchain/lex:tokenized_buffer",
         "@llvm-project//llvm:Support",
     ],
@@ -143,7 +144,7 @@ cc_test(
         "//common:ostream",
         "//testing/base:gtest_main",
         "//testing/base:test_raw_ostream",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:mocks",
         "//toolchain/lex",
@@ -164,7 +165,7 @@ cc_fuzz_test(
         ":parse",
         "//common:check",
         "//testing/fuzzing:libfuzzer_header",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/diagnostics:null_diagnostics",
         "//toolchain/lex",
