+++ toolchain/parse/parse_fuzzer.cpp
@@ -7,7 +7,7 @@
 
 #include "llvm/ADT/StringRef.h"
 #include "testing/fuzzing/libfuzzer.h"
-#include "toolchain/base/value_store.h"
+#include "toolchain/base/shared_value_stores.h"
 #include "toolchain/diagnostics/null_diagnostics.h"
 #include "toolchain/lex/lex.h"
 #include "toolchain/parse/parse.h"
