+++ toolchain/sem_ir/BUILD
@@ -36,7 +36,8 @@ cc_library(
         "//common:check",
         "//common:ostream",
         "//toolchain/base:index_base",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
+        "//toolchain/base:value_ids",
         "//toolchain/diagnostics:diagnostic_emitter",
         "//toolchain/parse:node_kind",
         "//toolchain/sem_ir:builtin_inst_kind",
@@ -75,6 +76,7 @@ cc_library(
         "//common:ostream",
         "//common:struct_reflection",
         "//toolchain/base:index_base",
+        "//toolchain/base:value_store",
         "@llvm-project//llvm:Support",
     ],
 )
@@ -127,6 +129,8 @@ cc_library(
         "//common:ostream",
         "//common:set",
         "//toolchain/base:kind_switch",
+        "//toolchain/base:shared_value_stores",
+        "//toolchain/base:value_ids",
         "//toolchain/base:value_store",
         "//toolchain/base:yaml",
         "//toolchain/lex:token_kind",
@@ -159,7 +163,7 @@ cc_library(
         ":inst_kind",
         "//common:ostream",
         "//toolchain/base:kind_switch",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/lex:tokenized_buffer",
         "//toolchain/parse:tree",
         "@llvm-project//llvm:Support",
@@ -177,7 +181,7 @@ cc_library(
         ":inst_namer",
         "//common:ostream",
         "//toolchain/base:kind_switch",
-        "//toolchain/base:value_store",
+        "//toolchain/base:shared_value_stores",
         "//toolchain/lex:tokenized_buffer",
         "//toolchain/parse:tree",
         "@llvm-project//llvm:Support",
